# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SNM API",
    "description": "API per il progetto \"Social Network for Music\"",
    "version": "1.0.0"
  },
  "host": "localhost:3004",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/spotify/getGenres": {
      "get": {
        "description": "Recupero dei generi musicali",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spotify/searchArtist": {
      "get": {
        "description": "Ricerca di un artista",
        "parameters": [
          {
            "name": "artist",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spotify/searchTracks": {
      "get": {
        "description": "Ricerca di una traccia",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spotify/getTracks/{track_id}": {
      "get": {
        "description": "Recupero dei dati di una traccia",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spotify/getTracksByArtist": {
      "get": {
        "description": "Recupero delle tracce relative ad un'artista",
        "parameters": [
          {
            "name": "artist_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spotify/getRecommendations": {
      "get": {
        "description": "Recupero delle tracce raccomandate per l'utente",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/spotify/getReccomendationsByGenre": {
      "get": {
        "description": "Recupero delle tracce raccomandate per l'utente in base ai generi musicali",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}